requirements.txt
mcp
mysql-connector-python
